+++ bot.js
@@ -15,11 +15,10 @@ for (const file of commandFiles) {
     bot.commands.set(command.name, command);
 }
 
-const cooldowns = new Discord.Collection();
 
 bot.once('ready', () => {
     console.log(`Bot is ready and working in ${bot.guilds.size} servers with ${bot.users.size} users! The bot is on version ${version}`);
-    bot.user.setActivity(`Asomataru v1.1.4!`);
+    bot.user.setActivity(`Asomataru v1.1.5!`);
     bot.user.setStatus('online', 'Made by The3D#9801')
     console.log('Ready!');
 });