+++ commands/birthday.js
@@ -1,6 +1,6 @@
 module.exports = {
     name: 'birthday',
-    execute: async (message, args) => {
+    execute: async (message) => {
        message.channel.send("Birthday is over with, proceed with everything else to offer")
     },
 };