+++ commands/invite.js
@@ -1,0 +1,7 @@
+module.exports = {
+    name:'invite',
+    description:'Invites Asomataru to your server!', 
+    execute (message) {
+    message.channel.send("Join the support server! https://discord.gg/vRPgqtb")
+    },
+};